[minor] test: refactor inheritance tests using trait
