+++ plugins/BEdita/Core/tests/TestCase/ORM/Inheritance/TableTest.php
@@ -15,9 +15,8 @@ namespace BEdita\Core\Test\TestCase\ORM\Inheritance;
 
 use BEdita\Core\ORM\Inheritance\AssociationCollection;
 use BEdita\Core\ORM\Inheritance\Query;
-use BEdita\Core\ORM\Inheritance\Table;
 use Cake\Datasource\EntityInterface;
-use Cake\ORM\Table as CakeTable;
+use Cake\ORM\Table;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 
@@ -28,46 +27,7 @@ use Cake\TestSuite\TestCase;
  */
 class TableTest extends TestCase
 {
-
-    /**
-     * Fixtures
-     *
-     * @var array
-     */
-    public $fixtures = [
-        'plugin.BEdita/Core.fake_animals',
-        'plugin.BEdita/Core.fake_mammals',
-        'plugin.BEdita/Core.fake_felines',
-        'plugin.BEdita/Core.fake_articles',
-    ];
-
-    /**
-     * Table FakeAnimals
-     *
-     * @var \BEdita\Core\ORM\Inheritance\Table
-     */
-    public $fakeAnimals;
-
-    /**
-     * Table FakeMammals
-     *
-     * @var \BEdita\Core\ORM\Inheritance\Table
-     */
-    public $fakeMammals;
-
-    /**
-     * Table FakeFelines
-     *
-     * @var \BEdita\Core\ORM\Inheritance\Table
-     */
-    public $fakeFelines;
-
-    /**
-     * Table options used for initialization
-     *
-     * @var array
-     */
-    protected $tableOptions = ['className' => Table::class];
+    use FakeAnimalsTrait;
 
     /**
      * setUp method
@@ -78,21 +38,7 @@ class TableTest extends TestCase
     {
         parent::setUp();
 
-        $this->fakeFelines = TableRegistry::get('FakeFelines', $this->tableOptions);
-        $this->fakeMammals = TableRegistry::get('FakeMammals', $this->tableOptions);
-        $this->fakeAnimals = TableRegistry::get('FakeAnimals');
-    }
-
-    /**
-     * Setup inheritance associations
-     *
-     * @return void
-     */
-    protected function setupAssociations()
-    {
-        $this->fakeMammals->extensionOf('FakeAnimals');
-        $this->fakeFelines->extensionOf('FakeMammals');
-        $this->fakeAnimals->hasMany('FakeArticles');
+        $this->setupTables();
     }
 
     /**
@@ -143,7 +89,7 @@ class TableTest extends TestCase
         $felinesInheritance = $this->fakeFelines->inheritedTable();
         static::assertEquals('FakeMammals', $felinesInheritance->getAlias());
 
-        $felinesDeepInheritance = array_map(function (CakeTable $inherited) {
+        $felinesDeepInheritance = array_map(function (Table $inherited) {
             return $inherited->getAlias();
         }, $this->fakeFelines->inheritedTables());
 
@@ -371,7 +317,7 @@ class TableTest extends TestCase
 
         $allColumns = $this->fakeFelines->getSchema()->columns();
         foreach ($this->fakeFelines->inheritedTables() as $t) {
-            if (!($t instanceof CakeTable)) {
+            if (!($t instanceof Table)) {
                 static::fail('Unexpected table object');
             }
